requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-Migrate==4.0.5
python-dotenv==1.0.1
pycups==2.0.1
# Para Windows, instale manualmente:
# pip install pywin32 Pillow
# Dependências adicionais recomendadas para produção e compatibilidade
python-engineio>=4.3.0
python-socketio>=5.7.2
eventlet>=0.33.0 # Para SocketIO em produção
# Para banco de dados SQLite (opcional, se usar outro, ajuste)
# sqlalchemy-utils
# Para suporte a .env
# python-dotenv
